+++ lib/src/http_client_none.dart
@@ -1,4 +1,7 @@
 import 'dart:async';
+import 'dart:typed_data';
+
+import 'package:swiss_knife/swiss_knife.dart';
 
 import 'http_client.dart';
 
@@ -22,3 +25,23 @@ HttpClientRequester createHttpClientRequesterImpl() {
 Uri getHttpClientRuntimeUriImpl() {
   return Uri(scheme: 'http', host: 'localhost', port: 80);
 }
+
+class HttpBlobNone extends HttpBlob {
+  HttpBlobNone(dynamic blob, MimeType mimeType) : super(blob, mimeType);
+
+  @override
+  int size() => 0;
+
+  @override
+  Future<ByteBuffer> readByteBuffer() {
+    throw UnimplementedError();
+  }
+}
+
+HttpBlob createHttpBlobImpl(dynamic content, MimeType mimeType) {
+  if (content == null) return null;
+  if (content is HttpBlob) return content;
+  return HttpBlobNone(content, mimeType);
+}
+
+bool isHttpBlobImpl(dynamic o) => o is HttpBlob;